moved saving of snapshots to own file
